Separate model conversion and loading
