+++ whisper/whisper/transcribe.py
@@ -40,20 +40,20 @@ def _format_timestamp(seconds: float):
 
 class ModelHolder:
     model = None
-    model_name_or_path = None
+    model_path = None
 
     @classmethod
-    def get_model(cls, model_name_or_path: str, dtype: mx.Dtype):
-        if cls.model is None or model_name_or_path != cls.model_name_or_path:
-            cls.model = load_model(model_name_or_path, dtype=dtype)
-            cls.model_name_or_path = model_name_or_path
+    def get_model(cls, model_path: str, dtype: mx.Dtype):
+        if cls.model is None or model_path != cls.model_path:
+            cls.model = load_model(model_path, dtype=dtype)
+            cls.model_path = model_path
         return cls.model
 
 
 def transcribe(
     audio: Union[str, np.ndarray, mx.array],
     *,
-    model_name_or_path: str = "tiny",
+    model_path: str = "mlx_models/tiny",
     verbose: Optional[bool] = None,
     temperature: Union[float, Tuple[float, ...]] = (0.0, 0.2, 0.4, 0.6, 0.8, 1.0),
     compression_ratio_threshold: Optional[float] = 2.4,
@@ -73,10 +73,8 @@ def transcribe(
     audio: Union[str, np.ndarray, mx.array]
         The path to the audio file to open, or the audio waveform
 
-    model_name_or_path: str
-        The Whisper model. Can be any of ["tiny", "base", "small", "medium", "large"],
-        or a local folder in MLX format.
-        Default is "tiny".
+    model_path: str
+        The path to the Whisper model that has been converted to MLX format.
 
     verbose: bool
         Whether to display the text being decoded to the console. If True, displays all the details,
@@ -116,7 +114,7 @@ def transcribe(
     """
 
     dtype = mx.float16 if decode_options.get("fp16", True) else mx.float32
-    model = ModelHolder.get_model(model_name_or_path, dtype)
+    model = ModelHolder.get_model(model_path, dtype)
 
     # Pad 30-seconds of silence to the input audio, for slicing
     mel = log_mel_spectrogram(audio, n_mels=model.dims.n_mels, padding=N_SAMPLES)
